api/requirements.txt
flask
python-dotenv
requests
xmltodict
pymongo